requirements.txt
# AI Framework & Orchestration
langchain
langchain-openai
langchain-community
langchain-pinecone
langsmith

# Clients & APIs
openai
pinecone-client

# Utilities
python-dotenv
tiktoken

# Data & UI
streamlit
numpy
pandas